+++ src/Cell.php
@@ -16,17 +16,16 @@ namespace MarcinOrlowski\AsciiTable;
 
 class Cell
 {
-    public function __construct(\Stringable|string|int $value,
-                                Align                  $align = Align::AUTO,
-                                Span|int               $columnSpan = Span::NONE)
+    public function __construct(\Stringable|string|int|float|bool|null $value,
+                                Align                                  $align = Align::AUTO)
     {
         $this->setValue($value);
-        $this->setColumnSpan($columnSpan);
         $this->setAlign($align);
     }
 
     /* ****************************************************************************************** */
 
+    /** String representation of cell's value. */
     protected string $value;
 
     public function getValue(): string
@@ -34,29 +33,21 @@ class Cell
         return $this->value;
     }
 
-    protected function setValue(\Stringable|string|int $value): self
+    protected function setValue(\Stringable|string|int|float|bool|null $value): self
     {
-        if ($value instanceof \Stringable) {
+        if ($value === null) {
+            $value = 'NULL';
+        } elseif ($value instanceof \Stringable) {
             $value = $value->__toString();
         } elseif (\is_int($value)) {
             $value = (string)$value;
+        } elseif (\is_bool($value)) {
+            $value = $value ? 'TRUE' : 'FALSE';
+        } elseif (\is_float($value)) {
+            $value = (string)$value;
         }
+
         $this->value = $value;
-        return $this;
-    }
-
-    /* ****************************************************************************************** */
-
-    protected Span|int $columnSpan = Span::NONE;
-
-    public function getColumnSpan(): Span|int
-    {
-        return $this->columnSpan;
-    }
-
-    protected function setColumnSpan(Span|int $columnSpan): self
-    {
-        $this->columnSpan = $columnSpan;
         return $this;
     }