+++ src/Row.php
@@ -14,13 +14,21 @@ declare(strict_types=1);
 
 namespace MarcinOrlowski\AsciiTable;
 
-use Traversable;
+use MarcinOrlowski\AsciiTable\Exceptions\DuplicateColumnKey;
+use MarcinOrlowski\AsciiTable\Traits\ArrayAccessTrait;
 
-class Row implements \ArrayAccess, \IteratorAggregate
+class Row implements ContainerContract
 {
+    use ArrayAccessTrait;
+
+    /**
+     * @param array|null $cells Optional list of cells to be added to the newly created row.
+     *
+     * @throws DuplicateColumnKey
+     */
     public function __construct(array|null $cells = null)
     {
-        $this->cells = new CellsContainer();
+        $this->container = new CellsContainer();
 
         if (!empty($cells)) {
             foreach ($cells as $columnKey => $cell) {
@@ -31,15 +39,17 @@ class Row implements \ArrayAccess, \IteratorAggregate
 
     /* ****************************************************************************************** */
 
-    protected CellsContainer $cells;
+    protected CellsContainer $container;
 
-    public function getCells(): CellsContainer
+    public function getContainer(): CellsContainer
     {
-        return $this->cells;
+        return $this->container;
     }
 
     /**
-     * @param array<string|int, Cell|string|int> $cells
+     * @param array<string|int, Cell|string|int|float|bool|null> $cells
+     *
+     * @throws DuplicateColumnKey
      */
     public function addCells(array $cells): self
     {
@@ -50,49 +60,25 @@ class Row implements \ArrayAccess, \IteratorAggregate
         return $this;
     }
 
+    /**
+     * @param string|int      $columnKey
+     * @param Cell|string|int $cell
+     * @param Align           $align
+     *
+     * @return $this
+     * @throws \MarcinOrlowski\AsciiTable\Exceptions\DuplicateColumnKey
+     */
     public function addCell(string|int      $columnKey,
                             Cell|string|int $cell,
-                            Align           $align = Align::AUTO,
-                            Span|int        $columnSpan = Span::NONE): self
+                            Align           $align = Align::AUTO): self
     {
         if (!($cell instanceof Cell)) {
-            $cell = new Cell($cell, $align, $columnSpan);
+            $cell = new Cell($cell, $align);
         }
-        $this->cells->add($columnKey, $cell);
+        $this->getContainer()->add($columnKey, $cell);
 
         return $this;
     }
 
     /* ****************************************************************************************** */
-
-    public function offsetExists(mixed $offset): bool
-    {
-        /** @var string|int $offset */
-        return $this->cells->offsetExists($offset);
-    }
-
-    public function offsetGet(mixed $offset): mixed
-    {
-        /** @var string|int $offset */
-        return $this->cells->offsetGet($offset);
-    }
-
-    public function offsetSet(mixed $offset, mixed $value): void
-    {
-        /** @var string|int $offset */
-        $this->cells->offsetSet($offset, $value);
-    }
-
-    public function offsetUnset(mixed $offset): void
-    {
-        /** @var string|int $offset */
-        $this->cells->offsetUnset($offset);
-    }
-
-    /* ****************************************************************************************** */
-
-    public function getIterator(): Traversable
-    {
-        return $this->cells->getIterator();
-    }
 }