+++ src/RowsContainer.php
@@ -14,65 +14,13 @@ declare(strict_types=1);
 
 namespace MarcinOrlowski\AsciiTable;
 
-use Traversable;
+use MarcinOrlowski\AsciiTable\Traits\ArrayAccessTrait;
 
 class RowsContainer implements \Countable, \IteratorAggregate, \ArrayAccess, ArrayableContract
 {
-    /** @var Row[] $rows */
-    protected array $rows = [];
+    use ArrayAccessTrait;
 
-    /* ****************************************************************************************** */
+    /** @var Row[] $container */
+    protected array $container = [];
 
-    public function count(): int
-    {
-        return \count($this->rows);
-    }
-
-    /* ****************************************************************************************** */
-
-    public function getIterator(): Traversable
-    {
-        return new \ArrayIterator($this->rows);
-    }
-
-    /* ****************************************************************************************** */
-
-    public function offsetExists(mixed $offset): bool
-    {
-        /** @var string|int $offset */
-        return \array_key_exists($offset, $this->rows);
-    }
-
-    public function offsetGet(mixed $offset): mixed
-    {
-        /** @var string|int $offset */
-        return $this->rows[ $offset ];
-    }
-
-    public function offsetSet(mixed $offset, mixed $value): void
-    {
-        if (!($value instanceof Row)) {
-            throw new \InvalidArgumentException('Invalid row type: ' . \get_debug_type($value));
-        }
-
-        if ($offset === null) {
-            $this->rows[] = $value;
-        } else {
-            /** @var string|int $offset */
-            $this->rows[ $offset ] = $value;
-        }
-    }
-
-    public function offsetUnset(mixed $offset): void
-    {
-        /** @var string|int $offset */
-        unset($this->rows[ $offset ]);
-    }
-
-    /* ****************************************************************************************** */
-
-    public function toArray(): array
-    {
-        return $this->rows;
-    }
 }