+++ CHANGES.md
@@ -6,6 +6,14 @@ Fast and flexible PHP library rendering plain ASCII tables.
 
 ## Changes
 
+* 1.3.0 (2022-10-18)
+  * Improved code documentation.
+  * Renamed `OutputContract` to `WriterContract`.
+  * Added own exceptions for most critical errors.
+  * Added support for individual cell alignment.
+  * Added more tests.
+  
+
 * v1.2.0 (2022-10-17)
   * Main class name changed to `AsciiTable`.
   * Added option to configure column default align.
@@ -15,9 +23,11 @@ Fast and flexible PHP library rendering plain ASCII tables.
   * Improved code creating row cells from plain string arrays.
   * Added more tests.
 
+
 * v1.1.0 (2022-10-17)
   * Corrected `\ArrayAccess` implementation.
   * Added more tests.
-  
+
+
 * v1.0.0 (2022-10-17)
   * Initial public release.